fix to compile warnings

+++ trunk/2d/rbf_interpolation.cxx
@@ -127,7 +127,7 @@ PetscErrorCode rbf_interpolation(Vec xi, Vec yi, Vec gi, Vec ei, Vec wi,
   ierr = VecAssemblyEnd(particle.i);CHKERRQ(ierr);
   ierr = VecGetArray(particle.i,&particle.il);CHKERRQ(ierr);
   for(i=0; i<particle.nilocal; i++) {
-    isort[i] = particle.il[i];
+    isort[i] = (int) particle.il[i];
   }
   ierr = ISCreateGeneral(PETSC_COMM_WORLD,particle.nilocal,isort,&isx);CHKERRQ(ierr);
   ierr = VecRestoreArray(particle.i,&particle.il);CHKERRQ(ierr);
@@ -325,3 +325,4 @@ PetscErrorCode rbf_interpolation(Vec xi, Vec yi, Vec gi, Vec ei, Vec wi,
 }
 
 #endif
+
